requirements.txt
python-dotenv==1.0.1
matplotlib==3.7.1